in pom reactvate latex plugin in the current version
I must be self critical: the config should have been done in a way 
that old version 1.8 works with the new manual. 
This would have been quite simple: all injections checked in, 
and in config removed goals inj and vrs. 

That simple. 
The way I did, the original manual does not compile.