In CommandExecutor.execute replaced File.lastModified() by a solution based on Files.getLastModifiedTime This is because File.lastModified() cannot distinguish between epoche time 0 and IO problems, whereas Files.getLastModifiedTime explicitly throws an IOException in case.
This weakness had to be eliminated, because in future, target files are not only created with converters fed with a time stamp, 
but in addition the resulting targets are endowed with the same time stamp as modification time. 
This consistency between inner time and outer time is crucial for future features. 

In class CommandExecutor
- added new method modTimeOrNull translating the IOException into a null timestamp. 
  This is used directly in method execute to check before execution 
  and indirectly via isUpdatedOrWarn to check after. 
- adapted methods execute and isUpdatedOrWarn and reworked also the apidoc. 
  Emphasis is to show that in any relevant case a warning is emitted. 
-