implemented new reproducibiity check based on metadata directly as before via latexmk
core aspect is LatexProcessor.runPdfInfo extracting CreationDate

 core aspect  is also CommandExecutor replacing
  private Optional<Long> timestampOpt; by 
  private Map<String, String> env;

replaced method setTimestamp(...) by methods 
- envReset() for doing without environment 
- envUtc() for setting timezone but no timestamp when a document is put under reproducibility check 
- envSetTimestamp() when a document is already under repro check. 

added to a method execute a parameter of type Map representing an environment. 
changex.xml: backlog: anlyze which commands with environment and which without. 

renamed the many commands execute for a better overview. 

core aspect is reimplementation of LatexProcessor.create 
allowing to run with timezone UTC when a document shall be put under repo check. 

Settings: added lots of infos in pdfinfo and exiftool. 

no longer set timestamp as a file property, so eliminated WFU04 and TexFileUtils.setModificationTime(...)

replaced Optional carrying a timestamp by environment in CommandExecutor. 
This is more general, as the case goes into where there is no timestamp but a timezone: 
when starting with reproducibility check.