Bugfix: Aux.Glo is no longer based on file extension glo but on aux. Originally, the idea was that a glossary is present, if and only if a *latex run creates a glo file. This may be true in many cases, but not in general, e.g. because an option nomain exists suppressing the glo file.
Instead, the mechanism uses the aux file and searches for pattern PATTERN_NEED_MAKEGLOSSARIES_RUN. 
The technique is similar as for bibliographies detected by the presence of the pattern PATTERN_NEED_BIBTEX_RUN. 

Consequently, the FileId shall not  be based on the glo file. 
Instead, the aux file is scanned for PATTERN_INPUT_GLOSSARY indicating a particular glossary. 
This pattern defines a group named GRP_EXT_GLOSS 
extracting the extension(s) of the files containing the entries of the various glossaries. 
Typically one of these extensions is really `glo`. 
The FileId shall be based on the files determined from the lines matching PATTERN_INPUT_GLOSSARY.