rework on endings related with synctex and a little bugfix mainly took ending synctex into accound besides synctex.gz This affects mainly documentation including figures. Also pdf files for comparisons.
In java class Settings besides synctex.gz also synctex was already taken into account, 
but not in according pom4pdf.xml 
There is, where a bugfix (in fact more than one) comes into the game: 
Added according entries in pom4pdf.xml but going beyond this including also depytx. 

Also added ending synctex to .latexmkrc. 
This was done in a modern way, @generated_exts instead of $clean_ext
In the course of this, it turned out, that also the Settings are not ok completely. 
So corrected this also.